Fixes #357 arrow key nav. in TOC
The up, down, end, and home arrow keys can be used to navigate the TOC rather than tab.
The TOC close button is left in the tab order but other items are navigated via arrow
keys. When a new item is loaded, focus remains in the TOC if it is open. Each time the TOC is
opened via the keyboard, focus starts on the TOC close button.