Fixes #488 Dialogs not working with screen readers
The settings and add to library dialogs were marked as
aria-hidden=true and the value was never changed
when the dialog was made visible.  Since the
visibility is controlled via CSS, the aria-hidden is not
needed and was removed.
Also removed the role=button on the input type=file elements
in add to library dialog. The role=button was preventing these
from working properly with the screen reader.