Fixes #252 Tooltip format on prev/next buttons
The prev/next page buttons were using the default browser title
tooltips rather than the bootstrap tooltips used on the other buttons.
Now all are the same and the prev/next button tooltips should be properly
placed. Note that I also increased the hideLoop timeout from 4 to 8
seconds to match the timer on the display of tooltips.